isPositiveNumber method changed

+++ utils/src/main/java/com/epam/utils/StringUtils.java
@@ -3,7 +3,7 @@ package com.epam.utils;
 public class StringUtils {
     public static boolean isPositiveNumber(String str) {
         try {
-            int num = Integer.parseInt(str);
+            double num = Double.parseDouble(str);
             return num > 0;
         } catch (NumberFormatException e) {
             return false;
